object-allocator: Allocate fixed size objects
The XDP2 object allocator allows defining zone allocators for various
sized objects. This is a a basic free list type allocator where the
objects are taken from a contiguous memory region. xdp2_obj_alloc_init
is the initialization function for an allocator that takes the number
of objects for the allocator and the size of each object as arguments.
The total size (num objects * object size) is malloc'ed and then an
initial free list of objects is created from the allocation.

Since the memory allocation for the allocator is contguous space, a
compressed "address" can be derived as the index of the object in
the block (e.g., index = (obj_address - base_address) / obj_size).
This designed assumes that the index fits into 16 bits hence an
allocator may have up to 64K objects.

xdp2_obj_alloc_alloc allocates an object. And returns the address
and index of the object. If there are no free objects then NULL
is returned

xdp2_obj_alloc_free frees an object

xdp2_obj_alloc_index_to_obj converts the index of an object to
its real address

xdp2_obj_alloc_obj_to_index converts the address of an object to
it index

There are also a number of sanity checks to validate the state
of the allocator and its objects

xdp2_obj_alloc_show_allocator is called to show information about
an object allocator inc;uding number of free objects and number of
allocations performed

A magic number is defined to check a pointer to an object allocator
passed to a obj_allocator function is indeed an object allocator.
A magic_num field is defined in struct xdp2_obj_allocator, and when
an object allocator  is initialized the field is set to the magic
value (0x43b3c9ef64bba98c). When an object allocator function is
subsequently called, an assert is one to check the magic number of
the input object

Create obj_alloc_free_list* functions for init, alloc, free, and
free_by_index

Object allocators are structures that potentially live in shared memory
so use the address translation facility with pointers of the object
allocator structure, particularly the base address and object pointers
ease enter the commit message for your changes. Lines starting
th '#' will be ignored, and an empty message aborts the commit.

Signed-off-by: Tom Herbert <tom@herbertland.com>