parser: XDP2 parser engine
Definitions and parsing engine for the XDP2 parser.

parser.c implements the main parsing loop in xdp2_parse as well as
protocol lookup functions with a parse node as input. xdp2_parse
calls __xdp2_parse that is the my parse loop. This function
walks the parse graph by invoking the length and next proto functions
in protocol nodes. For each parse node, the extract metadata and
handle protocol functions are called.

Canned protocol nodes are lib/xdp2/proto_nodes and
lib/xdp2/proto_nodes.mk contains a list of protocol node objects to
build.

Similarly, canned parsers are lib/xdp2/parsers and lib/xdp2/parser.mk
contains a list of parsers to build.

parser.c implements adds TLV handling in the main parsing loop for
those parse nodes (and protocol nodes) that contain TLVs. For a parse
node that contains TLVs (a TLVs parse node) the TLVs are processed
after extracting metadata from the protocol layer. TLV parsing occurs
in a loop in the xdp2_parser_tlvs functions. To parse the TLVs, the
TLV length and type functions in the TLVs parse node are invoked. For
each TLV the extract metadata and process TLV functions of individual
TLV parse nodes may be invoked.

Add support in parse engine to parse flag-fields. This includes a
function to lookup a flag-field in a flag-fields protocol table by
index of a flag in a flag-fields table, the main function to perform
flag-fields parser (xdp2_parse_flag_fields) and calling that function
from xdp2_parse when the parse node type is flag-fields

Signed-off-by: Tom Herbert <tom@herbertland.com>