locks: Add locks.h
Add locks.h file that contains macro definitions for mutexes and
conidtional variables. The macros are used in fifo.h to allow and
alternative set of locking primitives not based on pthread

There are two variants for locks: pthreads, and atomics

The variant can be selected at compile time by setting -D define to
XDP2_LOCKS_USE_PTHREADS or XDP2_LOCKS_USE_ATOMICS. If neither of
those are set then the variant can be set at runtime via the
XDP2_LOCKS_variant. The environment variable can be set to "pthreads"
or "atomics". If the environment variable is not set then the default
is atomics

Add macro XDP2_LOCKS_ATOMICS_MUTEX_LOCK_DEBUG that is implemented
in the lock backend as a loop doing trylock per the particular lock
type. If the debug count is hit in the loop then a warning is
printed that includes the function and location trying to lock,
the last locker, and the current locker. The debug count is set
by environment variable XDP2_LOCKS_debug_count

Also, add an environment variable XDP2_LOCKS_atomic_lock_sleep that
sets the number of microseconds to sleep in between atomic checks
of a lock

Signed-off-by: Tom Herbert <tom@herbertland.com>