switch: Add XDP2 switch and select facility macros
Definitions for the XDP2 switch facility are in xdp2/switch.h. This
include functions to compare two structures (pointers) for equality,
with a mask for a ternary compare, and a prefix match for some length

The XDP2 switch facility allows an rich syntax beyond a C switch statement
to allow case values to things other than ordinal types. Also, the
evaluation function can be for equality, a ternary match, or a prefix
match by prefix length

Here is an example that switches on an Ethernet header:

XDP2_SWITCH_START(&ehdr)
	XDP2_SWITCH_CASE_EQUAL(ehdr2) {
		printf("Hello world\n");
		break;
	}
	XDP2_SWITCH_CASE_EQUAL_CONST(({ 0x0, 0x11, 0x11, 0x11, 0x11, 0x11 },
				     { 0x0, 0x11, 0x11, 0x11, 0x11, 0x22 },
				     ETH_P_IP)) {
		printf("Match 1\n");
		break;
	}
	XDP2_SWITCH_CASE_TERN(ehdr2, ehdr2)
		printf("Match 2\n");
	XDP2_SWITCH_CASE_TERN_CONST(({ 0x0, 0x11, 0x11, 0x11, 0x11, 0x11 },
				    { 0x0, 0x11, 0x11, 0x11, 0x11, 0x22 },
				    ETH_P_IP),
				    ({ 0x0, 0x11, 0x11, 0x11, 0x11, 0x11 },
				    { 0x0, 0x11, 0x11, 0x11, 0x11, 0x22 },
				    ETH_P_IP)) break;
	XDP2_SWITCH_CASE_PREFIX(ehdr2, 23)
		break;
	XDP2_SWITCH_CASE_PREFIX_CONST(({ 0x0, 0x11, 0x11, 0x11, 0x11, 0x11 },
				      { 0x0, 0x11, 0x11, 0x11, 0x11, 0x22 },
				      ETH_P_IP), 43) {
		printf("Match 3\n");
		break;
	}
	XDP2_SWITCH_DEFAULT() printf("No hits\n");
XDP2_SWITCH_END();

The XDP2 select facility allows an N-way switch. Effectively, this would
give:

switch (a, b, c) {
	case: 1 ,2 ,3:
		...
}

Where a, b, c are the select types and have ordinal types.

XDP2_SELECT_CASE(value, ...) Is taken is all parameters are equal to
the targets. If a parameter is omitted then it is a wildcard

XDP2_SELECT_CASE_RANGE((high, low), ...) Each taget parameter is
compared to a range (inclusive endpoints). Either the high or
low values may be omitted s wildcard and are minimum or maximum
inifinity

XDP2_SELECT_CASE_MASK((value, mask), ..) Each target is compared by
target & mask == value & mask. If value or mask is omitted then the
case is a wildcard that matches all inputs

XDP2_SELECT_CASE_MASK_ANY(mask, ...) Each target is anded with a
mask and the result is checked to be non-zero. If a mask is omitted
then the case is a wildcard that matches all inputs

The XDP2_SELECT macros do this. Here is an example:

XDP2_SELECT_START(a, b, c)
	XDP2_SELECT_CASE(10, 1, 54)
		break;
	XDP2_SELECT_CASE(, , 54) { /* Omitted parameters are wildcards */
		break;
	}
	XDP2_SELECT_CASE(10, 25) /* Last parameter is omitted wildcard */
		printf("Part 2\n");
	}
	XDP2_SELECT_CASE_RANGE((1, 5), (10, 10), (20, 30)) {
		break;
	}
	XDP2_SELECT_CASE_RANGE((1, 5), (88,), (20, 30)) {
		/* Omitted top range is unlimited */
		break;
	}
	XDP2_SELECT_CASE_RANGE((1, 5), (,), (20, 30)) {
		/* Second parameter is wildcard rnage */
		break;
	}
	XDP2_SELECT_CASE_MASK((1, 5), (10, 0x33), (20, 0x3f)) {
		break;
	}
	XDP2_SELECT_CASE_MASK((1, 5), (,), (20, 0x3f)) {
		break;
	}
	XDP2_SELECT_CASE_MASK_ANY(0x1, 0xff, 2) {
		break;
	}
	XDP2_SELECT_CASE_MASK_ANY(0x1, , 2) {
		break;
	}
	XDP2_SELECT_DEFAULT() printf("No hits\n");
XDP2_SELECT_END();

Signed-off-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: Tom Herbert <tom@sipanda.io>