timers: Add XDP2 timers
This is a fairly simple timer wheel implementation. An instance of a
timer is defined by the timer_wheel structure. Like most timer wheel
implementations this contains an array of linked list timer structures

A timer wheel has no defined granularity. Granaularity is a property
applied by the user of the timer wheel. For instance, an application
might define several timer wheels for different granularities: like
nsces, usces, msecs, etc. To support this, there are two callbacks
set for a timer wheel:
    - get_current_time: Returns the current time in the units of
      granularity for the timer wheel. For instance, if the implied
      granularity is milliseconds then the time is returned in
      milliseconds
    - set_next_timer_pop: Set the next time that the application should
      call xdp2_run_timer

The timers are processed for expiration when xdp2_run_timer_wheel is
called. This can be driven by a pop_timer maintained by the application.
For instance, the pop timer could be a running interval timer or could
be a programmed hardware time. The timer_wheel implementation invokes the
set_next_timer_pop callback to indicate the next time that the timer wheel
should be run (that is the time of the next timer that will expire)

A timer is specified by the xdp2_timer structure. In the API, the
timer structure is provided by the caller. A timer structure should be
initialized to zeros, and the callback and argument fields should be
set by the caller

The API functions are:
    - xdp2_timer_create_wheel: Mallocs and initializes a timer wheel.
      Arguments include number of slots, the get_current_time and
      set_next_timer_pop functions
    - xdp2_timer_add: Add a timer with some delay time to fire ((relative
      to the current time, not absolute time). Arguments are the
      timer wheel, the timer structure, and the expiration time. If the
      timer is already running then it's removed and re-added
    - xdp2_timer_remove: Remove a timer from the timer wheel. Arguments
      are the timer wheel and the timer structure. Note that the timer
      structure must be the same one that was added
    - xdp2_run_timer_wheel: Run the timer wheel. Scan the timer wheel
      and for any expired timers (i.e. their expiration time is less than
      or equal to the current time), call the associated callback
      function and remove the timer from the timer wheel
    - xdp2_timer_show_wheel: Show the basic info and stats for a timer
      wheel
    - xdp2_timer_show_wheel_All: Show the basic info and stats for a timer
      wheel all the active timers

When a timer is removed there is still a possibility that the timer
may fire. Care must be taken to properly handle this especially if
the timer data structure is being freed.

1. If xdp2_timer_remove is called on a timer from within the timer's
   callback then the timer structure is free-able
2. If xdp2_timer_remove returns false (not called from a callback) then
   the timer structure is free-able
3. If xdp2_timer remove returns true when not called from the timer's
   callback then the timer structure is not-freeable. It is recommended
   that another add_timer be done with timeout of 0 delay. When the
   timer fires xdp2_timer_rmove should be call from the callback on
   the timer (just in case the callabck was for an old timeout), and
   then the structure is free-able per #1

Signed-off-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: Tom Herbert <tom@sipanda.io>