cli: Add CLI support in XDP2
Add helper functions an definitions as a frontend to libcli.

XDP2_CLI_ADD_SHOW_CONFIG is a helper macro to create a show
command

XDP2_CLI_ADD_SET_CONFIG is a helper macro to create a set
command

XDP2_CLI_SET_THREAD_INFO is a helper macro to create a thread info
structure that defines a CLI instance

xdp2_cli_start is a function to start a CLI. The input argument
is a thread info structure. When called a TCP listener is created
for the port number specified in the thread info structure, the
CLI is run in its own thread

Signed-off-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: Tom Herbert <tom@sipanda.io>