parser: Add array node parsing
Add array.h that defines protocol nodes and parse nodes for parsing
arrays. An "array parse node" is a top level parse node that contains
an array. An "arrel parse node" is a parse node for an array entry.
A xdp2_proto_array_def is the protocol definition structure for an
an array. There is no protocol definition for array elements since
they're so simple.